pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "shortcut-api"
version = "0.1.0"
description = "iOS Shortcut email alias + verification code extraction backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.29",
    "alembic>=1.13.1",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.1",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.29.0",
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.127.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.2.1",
    "slowapi>=0.1.9",
    "httpx>=0.27.0",
    "tenacity>=8.3.0",
    "shortuuid>=1.0.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "anyio>=4.3.0",
    "ruff>=0.4.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
